refactor: align lowering data with owning modules
